Refactor project into ford_fulkerson package
